Remove the c-parser.ts, for the antlr4 types. As this is not allowed in the antlr4-test.ts
